# ⚡ Session Caching + HAR Analysis: 25x Performance Boost