--- a/docs/solution/FINDINGS.md
+++ b/docs/solution/FINDINGS.md
@@ -1,0 +1,135 @@
+# HAR Analysis Findings & Recommendations
+
+## Key Discovery from HAR File Analysis
+
+### What We Found:
+1. **Real Chrome browser sends ZERO cookies** when accessing SET API endpoints
+2. **All requests succeed with HTTP 200** without any cookies
+3. **Our generated cookies are unnecessary** and may actually trigger bot detection
+
+### HAR Evidence:
+```
+🍪 Cookies: (empty)
+
+✅ 200 - AAI/highlight-data
+✅ 200 - 3BBIF/highlight-data
+✅ 200 - ADVANC/highlight-data
+✅ 200 - (all other endpoints)
+```
+
+### Test Results Comparison:
+
+| Test Scenario | Success Rate | Notes |
+|---------------|--------------|-------|
+| With complex generated cookies (initial) | 14% (14/100) | Many HTTP 452 errors |
+| With generated cookies (later) | 37% (37/100) | Better but inconsistent |
+| Chrome browser (HAR) | 100% (10/10) | No cookies sent |
+| Our service (13 stocks) | 100% (13/13) | Small sample works |
+
+## Root Cause Analysis
+
+The varying success rates (14% → 37% → 100%) indicate **rate limiting, not cookie issues**:
+
+1. **Rate Limiting**: Making 100 requests in rapid succession triggers Incapsula's rate limits
+2. **IP-Based Blocking**: Repeated failed requests from same IP get temporary blocks
+3. **Browser Fingerprinting**: curl_cffi's Chrome impersonation is what matters, not cookies
+
+## What Actually Matters for Bot Detection Bypass
+
+### ✅ Critical Factors (Working):
+1. **Browser Impersonation**: curl_cffi with `impersonate="chrome120"`
+2. **Proper Headers**:
+   - User-Agent: Chrome 140
+   - Sec-CH-UA headers
+   - Referer: Symbol-specific URL
+   - Accept headers
+   - Cache-Control, Pragma
+3. **Request Timing**: Avoid rapid-fire requests
+
+### ❌ Not Critical (Can Remove):
+1. **Cookie Generation**: No cookies needed!
+2. **Analytics Cookies**: _ga, _fbp, _gcl_au - unused
+3. **Session Cookies**: __lt__sid, _hjSession - unused
+4. **Incapsula Cookies**: visid_incap, incap_ses - not required
+
+## Recommendations
+
+### 1. Remove Cookie Generation (High Priority)
+```python
+# Current (unnecessary complexity):
+cookies = (
+    self.session_cookies
+    or AsyncDataFetcher.generate_incapsula_cookies(landing_url=referer)
+)
+
+# Recommended (simpler, matches real browser):
+cookies = self.session_cookies or None  # Only use if provided by user
+```
+
+### 2. Add Rate Limiting Protection
+```python
+# Add delay between requests in services
+import asyncio
+await asyncio.sleep(random.uniform(0.1, 0.3))  # 100-300ms delay
+```
+
+### 3. Keep What Works
+- ✅ curl_cffi browser impersonation
+- ✅ Symbol-specific referer headers
+- ✅ Proper Chrome 140 headers
+- ✅ Accept real browser cookies if user provides them
+
+### 4. Production Strategy
+
+**For Development/Testing:**
+- Use the library as-is with generated cookies (works for small batches)
+- Add delays between requests to avoid rate limiting
+
+**For Production:**
+- **Option A**: Use real browser session cookies (100% reliable)
+  - Capture cookies from actual Chrome session
+  - Refresh periodically (every few hours)
+
+- **Option B**: Disable cookie generation entirely
+  - Remove generated cookies
+  - Rely on headers + browser impersonation only
+  - Add rate limiting (1-2 requests/second max)
+
+- **Option C**: Use rotating residential proxies
+  - Distribute requests across multiple IPs
+  - Avoid rate limiting per IP
+
+## Code Changes Needed
+
+### Priority 1: Simplify Cookie Handling
+File: `settfex/utils/data_fetcher.py`
+- Keep `generate_incapsula_cookies()` for backwards compatibility
+- Add note that cookies are optional
+- Document that Chrome sends no cookies by default
+
+### Priority 2: Add Rate Limit Protection
+File: `settfex/services/set/stock/*.py`
+- Add optional delay parameter
+- Implement smart rate limiting (exponential backoff on 452 errors)
+
+### Priority 3: Update Documentation
+- Explain HAR findings
+- Show that cookies are optional
+- Provide production deployment strategies
+
+## Conclusion
+
+**The generated cookies we spent so much effort perfecting are actually not needed!**
+
+Real Chrome succeeds with:
+- ✅ No cookies
+- ✅ Proper headers
+- ✅ Browser impersonation
+- ✅ Reasonable request rate
+
+Our failures are due to:
+- ❌ Rate limiting from rapid requests
+- ❌ Possibly IP-based blocking from failed attempts
+- ❌ NOT missing or incorrect cookies
+
+The simplest solution: Remove cookie generation, add rate limiting, keep the excellent header configuration we already have.