--- a/pyproject.toml
+++ b/pyproject.toml
@@ -8,6 +8,7 @@ dependencies = [
     "curl-cffi>=0.6.0",
     "pydantic>=2.0.0",
     "loguru>=0.7.0",
+    "diskcache>=5.6.0",
 ]
 authors = [
     { name = "Your Name", email = "your.email@example.com" }