--- a/settfex/services/set/stock/stock.py
+++ b/settfex/services/set/stock/stock.py
@@ -32,6 +32,7 @@ class Stock:
         symbol: str,
         config: FetcherConfig | None = None,
         session_cookies: str | None = None,
+        use_cookies: bool = True,
     ) -> None:
         """
         Initialize Stock instance for a specific symbol.
@@ -40,25 +41,32 @@ class Stock:
             symbol: Stock symbol (e.g., "CPALL", "PTT", "kbank")
             config: Optional fetcher configuration
             session_cookies: Optional browser session cookies
+            use_cookies: Whether to generate cookies automatically. Default True (recommended
+                        for compatibility). Set to False only if providing session_cookies.
 
         Example:
-            >>> # Basic usage
+            >>> # Basic usage (auto-generated cookies - recommended)
             >>> stock = Stock("CPALL")
             >>>
-            >>> # With custom config
-            >>> config = FetcherConfig(timeout=60, max_retries=5)
+            >>> # With rate limiting to avoid HTTP 452
+            >>> config = FetcherConfig(rate_limit_delay=0.2)
             >>> stock = Stock("CPALL", config=config)
             >>>
-            >>> # With session cookies
+            >>> # With real browser session cookies (most reliable)
             >>> cookies = "charlot=abc123; incap_ses_357_2046605=xyz789; ..."
             >>> stock = Stock("CPALL", session_cookies=cookies)
+            >>>
+            >>> # No cookies mode (may get HTTP 403)
+            >>> stock = Stock("CPALL", use_cookies=False)
         """
         self.symbol = normalize_symbol(symbol)
         self.config = config
         self.session_cookies = session_cookies
+        self.use_cookies = use_cookies
 
         # Initialize service instances (lazy initialization for future services)
         self._highlight_data_service: StockHighlightDataService | None = None
+        self._profile_service: StockProfileService | None = None
 
         logger.info(f"Stock instance created for symbol '{self.symbol}'")
 
@@ -72,7 +80,9 @@ class Stock:
         """
         if self._highlight_data_service is None:
             self._highlight_data_service = StockHighlightDataService(
-                config=self.config, session_cookies=self.session_cookies
+                config=self.config,
+                session_cookies=self.session_cookies,
+                use_cookies=self.use_cookies,
             )
         return self._highlight_data_service
 
@@ -103,6 +113,47 @@ class Stock:
             symbol=self.symbol, lang=lang
         )
 
+    @property
+    def profile_service(self) -> "StockProfileService":
+        """
+        Get or create profile service instance.
+
+        Returns:
+            StockProfileService instance
+        """
+        if self._profile_service is None:
+            from settfex.services.set.stock.profile_stock import StockProfileService
+
+            self._profile_service = StockProfileService(
+                config=self.config,
+                session_cookies=self.session_cookies,
+                use_cookies=self.use_cookies,
+            )
+        return self._profile_service
+
+    async def get_profile(self, lang: str = "en") -> "StockProfile":
+        """
+        Fetch profile data for this stock.
+
+        Args:
+            lang: Language for response ('en' or 'th', default: 'en')
+
+        Returns:
+            StockProfile with company and listing information
+
+        Raises:
+            ValueError: If language is invalid
+            Exception: If request fails
+
+        Example:
+            >>> stock = Stock("PTT")
+            >>> profile = await stock.get_profile()
+            >>> print(f"Company: {profile.name}")
+            >>> print(f"Sector: {profile.sector_name}")
+        """
+        logger.debug(f"Fetching profile for {self.symbol} (lang={lang})")
+        return await self.profile_service.fetch_profile(symbol=self.symbol, lang=lang)
+
     # Future service methods (placeholders for documentation)
     # async def get_shareholders(self, lang: str = "en") -> ShareholdersData:
     #     """Fetch shareholder information for this stock."""
@@ -110,10 +161,6 @@ class Stock:
     #
     # async def get_financials(self, lang: str = "en") -> FinancialsData:
     #     """Fetch financial statements for this stock."""
-    #     pass
-    #
-    # async def get_company_profile(self, lang: str = "en") -> CompanyProfile:
-    #     """Fetch company profile information."""
     #     pass
 
     def __repr__(self) -> str: