… docs/session_caching.md
… docs/solution/SESSION_CACHE_SUMMARY.md
… docs/solution/SOLUTION_100_PERCENT.md
… settfex/utils/data_fetcher.py
… settfex/utils/session_cache.py
… uv.lock